200820 프로젝트4 완성
뷰페이저를 담아두었던 프래그먼트를 통째로 제거.
이를 메인 액티비티에서 구현.
전역변수를 이용해 충돌하는 메서드를 없애주었는데, 아예 메서드 자체를 간소화시켜, 전역 변수 사용할 필요가 없어짐.
프레임 레이아웃의 기본 상태가 뷰페이저이게 구현했고 상세보기 클릭시 이를 위에 얹고 또 제거하는 방식으로 코드를 간소화.
프래그먼트1에서 액티비티의 메서드 호출.
