+++ 8월/MyDetailPart/app/src/main/res/layout/activity_main.xml
@@ -32,7 +32,16 @@
             android:id="@+id/container"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_weight="4">
+            android:layout_weight="4"
+            android:background="@drawable/main_background">
+
+            <androidx.viewpager.widget.ViewPager
+                android:id="@+id/pager"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+            </androidx.viewpager.widget.ViewPager>
+
         </FrameLayout>
 
     </LinearLayout>
