+++ 8월/MyDetailPart/app/src/main/res/layout/fragment2.xml
@@ -2,8 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:background="@drawable/main_background">
+    android:orientation="vertical">
 
     <LinearLayout
         android:layout_width="match_parent"
